added the correct images (width and height + under 200kb and 72dpi)
added all the images on the home page again, cause i forgot to change the dpi on them to 72. now that is done, i also added css for the "program" making it responsive for all sizes. added some new <div classes> to make it easer to not repeat my css.