added all the content to exhibition page
wrote all the html i need on the exhibition page (might do some slight adjustments later) added pictures, all pictures are under 200kb and a dip of 72.
added footer to the exhibition page.
also added a @media to hide the search bar in the header when the screen is smaller than 1080px